drm/i915: New ioctl for getting available crtc
Add new i915 IOCTL to get available crtc for a given connector.
the IOCTL returns valid crtc id if successful, otherwise returns
negative error code.

For a connector that has a mode that needs joiner, check for available
pipe from reverse order and make sure that the next pipe is also
available that can be made as secondary.

Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>